Refactor FaucetForm
