+++ src/components/icons/InfoIcon.tsx
@@ -1,0 +1,12 @@
+"use client";
+import React from "react";
+import { createSvgIcon } from "@mui/material";
+
+export const InfoIcon = createSvgIcon(
+  <path
+    d="M2.25 12a9.75 9.75 0 1 1 19.5 0 9.75 9.75 0 0 1-19.5 0zm8.7-1.44a1.5 1.5 0 0 1 2.13 1.7l-.7 2.84.03-.02c.37-.17.8-.01.98.35.18.36.05.8-.3 1l-.04.01a1.5 1.5 0 0 1-2.13-1.7l.7-2.84-.03.02a.75.75 0 1 1-.68-1.34zM12 9a.75.75 0 1 0 0-1.5.75.75 0 0 0 0 1.5zm0 0"
+    fillRule="evenodd"
+    fill="currentColor"
+  />,
+  "Info"
+);
