+++ src/utils/makeCssVars.ts
@@ -1,0 +1,5 @@
+export function makeCssVars(
+  vars: Record<`--${string}`, string | number>
+): React.CSSProperties {
+  return vars as React.CSSProperties;
+}
